requirements.txt
feedparser>=6.0.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
newspaper3k>=0.2.8
google-generativeai>=0.4.0
python-dotenv>=1.0.0
rich>=13.0.0
lxml>=5.0.0
lxml_html_clean>=0.1.0
streamlit>=1.30.0

